# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AddFavoriteRequest:
    properties:
      product_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - product_id
    type: object
...
  handlers.FavoriteResponse:
    properties:
      favorited_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/handlers.ProductResponse'
      product_id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  handlers.FavoritesListResponse:
    properties:
      favorites:
        items:
          $ref: '#/definitions/handlers.FavoriteResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  handlers.ProductResponse:
    description: Product response with enhanced information
    properties:
      average_rating:
        type: number
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      comment_count:
        type: integer
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comments_count:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      curated_price:
        type: number
      curated_tags:
        items:
          type: string
        type: array
      description:
        type: string
      favorite_count:
        description: Additional fields from products handler
        type: integer
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      favorites_count:
        type: integer
      feature_vector:
        $ref: '#/definitions/models.ProductFeatureVector'
      has_suggestions:
        type: boolean
      id:
        type: string
      is_favorited:
        type: boolean
      name:
        type: string
      price:
        type: number
      product_suggestions:
        items:
          $ref: '#/definitions/models.ProductSuggestion'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      recommended_reason:
        type: string
      similar_products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      similarity_data_as_product:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      similarity_data_as_similar:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      tags:
        items:
          type: string
        type: array
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
...
  handlers.StandardErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.Category:
    properties:
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Comment:
    properties:
      body:
        type: string
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      created_at:
        type: string
      downvotes:
        type: integer
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      sentiment_label:
        type: string
      sentiment_score:
        type: number
      upvotes:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.CommentVote:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      comment_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      vote_type:
        type: string
    type: object
  models.Favorite:
    properties:
      favorited_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.HttpRequestLog:
    properties:
      attack_score:
        type: number
      duration_ms:
        type: integer
      id:
        type: string
      ip_address:
        type: string
      method:
        type: string
      path:
        type: string
      referrer:
        type: string
      status_code:
        type: integer
      suspected_attack_type:
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      curated_price:
        type: number
      curated_tags:
        items:
          type: string
        type: array
      description:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      feature_vector:
        $ref: '#/definitions/models.ProductFeatureVector'
      id:
        type: string
      name:
        type: string
      price:
        type: number
      product_suggestions:
        items:
          $ref: '#/definitions/models.ProductSuggestion'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      similarity_data_as_product:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      similarity_data_as_similar:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      tags:
        items:
          type: string
        type: array
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
  models.ProductFeatureVector:
    properties:
      embedding:
        items:
          type: number
        type: array
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ProductSimilarityData:
    properties:
      based_on:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      similar_product:
        $ref: '#/definitions/models.Product'
      similar_product_id:
        type: string
      similarity_score:
        type: number
    type: object
  models.ProductSuggestion:
    properties:
      generated_at:
        type: string
      id:
        type: string
      model_version:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      reason:
        type: string
      suggested_price_max:
        type: number
      suggested_price_min:
        type: number
      suggested_tags:
        items:
          type: string
        type: array
    type: object
  models.Recommendation:
    properties:
      age_based:
        type: boolean
      based_on_comments:
        type: boolean
      based_on_favorites:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      model_version:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      reason:
        type: string
      region_based:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.User:
    properties:
      birth_year:
        type: integer
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      http_request_logs:
        items:
          $ref: '#/definitions/models.HttpRequestLog'
        type: array
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      region:
        type: string
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
  models.UserEvent:
    properties:
      event_type:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
...
  /favorites:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of user's favorite products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Favorites retrieved successfully
          schema:
            $ref: '#/definitions/handlers.FavoritesListResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's favorites
      tags:
      - Favorites
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's favorites
      parameters:
      - description: Product to add to favorites
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddFavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product added to favorites successfully
          schema:
            $ref: '#/definitions/models.Favorite'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "409":
          description: Product already in favorites
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Add product to favorites
      tags:
      - Favorites
  /favorites/{product_id}:
    delete:
      consumes:
      - application/json
      description: Remove a product from the authenticated user's favorites
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Product removed from favorites successfully
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Favorite not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove product from favorites
      tags:
      - Favorites
  /favorites/{product_id}/check:
    get:
      consumes:
      - application/json
      description: Check if a specific product is in the authenticated user's favorites
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Favorite status
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Check if product is favorited
      tags:
      - Favorites
  /favorites/{product_id}/toggle:
    post:
      consumes:
      - application/json
      description: Add or remove a product from favorites based on current status
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Toggle result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle favorite status
      tags:
      - Favorites
  /favorites/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about the authenticated user's favorites
      produces:
      - application/json
      responses:
        "200":
          description: Favorite statistics
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get favorite statistics
      tags:
      - Favorites
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...